+++ src/attributecode/cmd.py
@@ -37,6 +37,7 @@ from attributecode import severities
 from attributecode.util import extract_zip
 from attributecode.util import update_severity_level_about_resource_path_not_exist_error
 from attributecode.util import to_posix
+from attributecode.util import inventory_filter
 
 
 __copyright__ = """
@@ -90,7 +91,7 @@ Generate licensing attribution and credit notices from .ABOUT files and inventor
 
 Read, write and collect provenance and license inventories from .ABOUT files to and from JSON or CSV files.
 
-Use about-code <command> --help for help on a command.
+Use about <command> --help for help on a command.
     """
 
 
@@ -108,6 +109,9 @@ Use about-code <command> --help for help on a command.
 @click.argument('output', nargs=1, required=True,
     type=click.Path(exists=False, dir_okay=False, resolve_path=True))
 
+@click.option('--filter', nargs=1, multiple=True, 
+    help='Filter for the output inventory.')
+
 @click.option('-f', '--format', is_flag=False, default='csv', show_default=True,
     type=click.Choice(['json', 'csv']),
     help='Set OUTPUT inventory file format.')
@@ -118,6 +122,10 @@ Use about-code <command> --help for help on a command.
 @click.option('--mapping-file', metavar='FILE', nargs=1,
     type=click.Path(exists=True, dir_okay=True, readable=True, resolve_path=True),
     help='Use a custom mapping file with mapping between input keys and ABOUT field names.')
+
+@click.option('--mapping-output', metavar='FILE', nargs=1,
+    type=click.Path(exists=True, dir_okay=True, readable=True, resolve_path=True),
+    help='Use a custom mapping file with mapping between ABOUT field names and output keys')
 
 @click.option('--verbose', is_flag=True, default=False,
     help='Show all errors and warnings. '
@@ -132,7 +140,7 @@ Use about-code <command> --help for help on a command.
 
 @click.help_option('-h', '--help')
 
-def inventory(location, output, mapping, mapping_file, quiet, format, verbose):
+def inventory(location, output, mapping, mapping_file, mapping_output, filter, quiet, format, verbose):
     """
 Collect a JSON or CSV inventory of components from .ABOUT files.
 
@@ -157,6 +165,22 @@ OUTPUT: Path to the JSON or CSV inventory file to create.
 
     errors, abouts = model.collect_inventory(location, use_mapping=mapping, mapping_file=mapping_file)
 
+    updated_abouts = []
+    if filter:
+        filter_dict = {}
+        # Parse the filter and save to the filter dictionary with a list of value
+        for element in filter:
+            key = element.partition('=')[0]
+            value = element.partition('=')[2]
+            if key in filter_dict:
+                filter_dict[key].append(value)
+            else:
+                value_list = [value]
+                filter_dict[key] = value_list
+        updated_abouts = inventory_filter(abouts, filter_dict)
+    else:
+        updated_abouts = abouts
+
     # Do not write the output if one of the ABOUT files has duplicated key names
     dup_error_msg = u'Duplicated key name(s)'
     halt_output = False
@@ -166,7 +190,7 @@ OUTPUT: Path to the JSON or CSV inventory file to create.
             break
 
     if not halt_output:
-        write_errors = model.write_output(abouts, output, format)
+        write_errors = model.write_output(updated_abouts, output, format, mapping_output)
         for err in write_errors:
             errors.append(err)
     else:
@@ -290,6 +314,12 @@ OUTPUT: Path to a directory where ABOUT files are generated.
     type=click.Path(exists=True, dir_okay=True, readable=True, resolve_path=True),
     help='Use a custom mapping file with mapping between input keys and ABOUT field names.')
 
+@click.option('--template', type=click.Path(exists=True), nargs=1,
+    help='Path to an optional custom attribution template used for generation.')
+
+@click.option('--vartext', nargs=1, multiple=True, 
+    help='Variable texts to the attribution template.')
+
 @click.option('--verbose', is_flag=True, default=False,
     help='Show all errors and warnings. '
         'By default, the tool only prints these '
@@ -298,15 +328,12 @@ OUTPUT: Path to a directory where ABOUT files are generated.
         'for any level.'
 )
 
-@click.option('--template', type=click.Path(exists=True), nargs=1,
-    help='Path to an optional custom attribution template used for generation.')
-
 @click.option('-q', '--quiet', is_flag=True,
     help='Do not print error or warning messages.')
 
 @click.help_option('-h', '--help')
 
-def attrib(location, output, template, mapping, mapping_file, inventory, quiet, verbose):
+def attrib(location, output, template, mapping, mapping_file, inventory, vartext, quiet, verbose):
     """
 Generate an attribution document at OUTPUT using .ABOUT files at LOCATION.
 
@@ -325,7 +352,7 @@ OUTPUT: Path to output file to write the attribution to.
     no_match_errors = attrib_generate_and_save(
         abouts=abouts, output_location=output,
         use_mapping=mapping, mapping_file=mapping_file, template_loc=template,
-        inventory_location=inventory)
+        inventory_location=inventory, vartext=vartext)
 
     if not no_match_errors:
         # Check for template error
@@ -378,21 +405,24 @@ LOCATION: Path to a .ABOUT file or a directory containing .ABOUT files.
 
     msg_format = '%(sever)s: %(message)s'
     print_errors = []
+    number_of_errors = 0
     for severity, message in errors:
         sever = severities[severity]
+        # Only problematic_errors should be counted.
+        # Others such as INFO should not be counted as error.
+        if sever in problematic_errors:
+            number_of_errors = number_of_errors + 1
         if verbose:
             print_errors.append(msg_format % locals())
         elif sever in problematic_errors:
             print_errors.append(msg_format % locals())
-
-    number_of_errors = len(print_errors)
 
     for err in print_errors:
         print(err)
 
     if print_errors:
         click.echo('Found {} errors.'.format(number_of_errors))
-        # FIXME: not sure this is the right way to exit with a retrun code
+        # FIXME: not sure this is the right way to exit with a return code
         sys.exit(1)
     else:
         click.echo('No error found.')